# Further improvements in runtime metadata loading