--- a/pkg/tfbridge/metadata.go
+++ b/pkg/tfbridge/metadata.go
@@ -61,3 +61,12 @@ func (info *MetadataInfo) assertValid() {
 	contract.Assertf(info.Path != "", "Path must be non-empty")
 
 }
+
+// trim the metadata to just the keys required for the runtime phase
+// in the future this method might also substitute compressed contents within some keys
+func (info *MetadataInfo) ExtractRuntimeMetadata() *MetadataInfo {
+	data, _ := metadata.New(nil)
+	metadata.CloneKey("auto-aliasing", info.Data, data)
+	metadata.CloneKey("mux", info.Data, data)
+	return &MetadataInfo{"runtime-bridge-metadata.json", ProviderMetadata(data)}
+}