--- a/pkg/tfbridge/metadata.go
+++ b/pkg/tfbridge/metadata.go
@@ -66,7 +66,7 @@ func (info *MetadataInfo) assertValid() {
 // in the future this method might also substitute compressed contents within some keys
 func (info *MetadataInfo) ExtractRuntimeMetadata() *MetadataInfo {
 	data, _ := metadata.New(nil)
-	metadata.CloneKey("auto-aliasing", info.Data, data)
+	metadata.CloneKey(aliasMetadataKey, info.Data, data)
 	metadata.CloneKey("mux", info.Data, data)
 	return &MetadataInfo{"runtime-bridge-metadata.json", ProviderMetadata(data)}
 }