--- a/pkg/tfgen/generate.go
+++ b/pkg/tfgen/generate.go
@@ -946,7 +946,11 @@ func (g *Generator) UnstableGenerateFromSchema(genSchemaResult *GenerateSchemaRe
 		files = map[string][]byte{"schema.json": bytes}
 
 		if info := g.info.MetadataInfo; info != nil {
-			files[info.Path] = (*metadata.Data)(info.Data).Marshal()
+			files[info.Path] = (*metadata.Data)(info.Data).MarshalIndent()
+			if true {
+				runtimeInfo := info.ExtractRuntimeMetadata()
+				files[runtimeInfo.Path] = (*metadata.Data)(runtimeInfo.Data).Marshal()
+			}
 		}
 	case PCL:
 		if g.skipExamples {