--- a/pkg/tfgen/generate.go
+++ b/pkg/tfgen/generate.go
@@ -947,7 +947,7 @@ func (g *Generator) UnstableGenerateFromSchema(genSchemaResult *GenerateSchemaRe
 
 		if info := g.info.MetadataInfo; info != nil {
 			files[info.Path] = (*metadata.Data)(info.Data).MarshalIndent()
-			if true {
+			if g.info.GenerateRuntimeMetadata {
 				runtimeInfo := info.ExtractRuntimeMetadata()
 				files[runtimeInfo.Path] = (*metadata.Data)(runtimeInfo.Data).Marshal()
 			}