Two changes to further reduce metadata load times for large providers:
1) Use segmentio/encoding/json in zero-copy mode
2) Generate a trimmed, runtime-only copy of metadata

In particular, removing extra whitespace contributed the most improvement suggesting that we may be able to obtain further savings just by shortening or eliminating frequently repeated strings.

This change is equivalent to the first codegen option described in https://github.com/pulumi/pulumi-aws/issues/2799#issuecomment-1802540034 (and combining the two yields no further improvement).

Generation of the runtime-only metadata is controlled via a flag in the ProviderInfo to allow providers to opt in individually. 


### Benchmarks from AWS:
Before changes
```
BenchmarkProviderGen
BenchmarkProviderGen-12             	       6	 189483917 ns/op	331080480 B/op	 2222893 allocs/op
BenchmarkProviderNoGen
BenchmarkProviderNoGen-12           	       7	 160893363 ns/op	287969720 B/op	 1850232 allocs/op
```

With zero-copy deserializer
```
BenchmarkProviderGen
BenchmarkProviderGen-12             	       6	 179858674 ns/op	331078597 B/op	 2222880 allocs/op
BenchmarkProviderNoGen
BenchmarkProviderNoGen-12           	       7	 155430929 ns/op	287963872 B/op	 1850211 allocs/op
```

With zero-copy deserializer and trimmed runtime only metadata
```
BenchmarkRuntimeProviderGen
BenchmarkRuntimeProviderGen-12      	       8	 132763682 ns/op	331080280 B/op	 2222892 allocs/op
BenchmarkRuntimeProviderNoGen
BenchmarkRuntimeProviderNoGen-12    	      10	 108300512 ns/op	287966743 B/op	 1850214 allocs/op
```